beaker: dn: centostream: Introducing centostream support [TEMPORARY]
It seems CentOStream is not yet supported on upstream Beaker.

Since we only care about the latest available version of centos8/9, and
will always import it from an open external repository, this is an ugly
hard-coded temporary workaround that changes how the repo of the upstream
CentOS distro is imported. Moreover, this is only called when the distro
link of the image has an url with the wildcard %centos%, such as:

* http://mirror.centos.org/centos/8-stream
* http://mirror.stream.centos.org/9-stream

Moreover, soft links to the harness RHEL8, RHEL9 packages are also
necessary for this workaround:

$ ln -s /var/www/beaker/harness/RedHatEnterpriseLinux8 /var/www/beaker/harness/CentOSStream8
$ ln -s /var/www/beaker/harness/RedHatEnterpriseLinux9 /var/www/beaker/harness/CentOSStream9

This will be removed in the future.

Signed-off-by: Desnes Nunes <desnesn@br.ibm.com>